Add files via upload
Why Encryption is important?
For maintaining the security of financial and other R&D data with various internal data which needs to be safe from intruders.
The employees or organizations do not want their financial or private items to be made public.
To make quick transitions so that private items are not made public.





Algorithm for file encryption:
1. Python accepts the file input and encrypts it using the Pycrypto module.
2. The filename is taken as input parameter along with the password.
3. Encryption is achieved with the help of key which is generated with SHA-256 algorithmic standards. The encrypted file is saved in the same directory with a prefix of (encrypted) added to it.
4. SHA-256 cryptographic hash function generates keys which helps in maintaining security of files that are used for symmetric encryption of files.
5. The files are basically protected with passwords.
6. Decryption follows the reverse procedure where the password and encrypted file is taken input parameters.


SUMMARY:

Encryption of images or file format is important if it includes any confidential information. The standard encryption technique which was followed earlier includes a lack of security which is achieved in the “PyAesCrypt” module and the demonstration displayed was for “.txt” messages but can be applied for any image format also.

Python is a highly popular language among programmers and is just behind Rust in terms of most loved programming languages. The simplicity of the language and shorter codes are some of the key reasons for its popularity. The jobs for Python developers are also growing and salaries can reach as high as $110,000 per year. The language is also handy when it comes to AI, Machine Learning and data science.

You can try learning more about Python through E-Books or books. Or you can also try out some of our tutorials pertaining to Python such as the “Python Programming An Expert Guide on Python” which comes with over 9 hours of video and includes topics such as Strings and list data types, variables, closures, functions, files, loops and much more.

Also, if you are familiar with the essentials of Python programming, you can go for the “Beyond Basic Programming – Intermediate Python” which comes over 3 hours of video and includes 7 sections such as cross-cutting tools, intermediate programming, functional programming, applications and much more.